freestanding: string: Have digit as a 64 bit wide integer in `itoa_z`
Fixes a compiler warning when compiling with GCC:

freestanding/string/string.c: In function 'itoa_z':
freestanding/string/string.c:165:38: warning: conversion from 'u64' {aka 'long long unsigned int'} to 'unsigned int' may change value [-Wconversion]
  165 |                 unsigned int digit = value % (unsigned int)base;

+++ freestanding/string/string.c
@@ -162,8 +162,8 @@ void itoa_z(u64 value, char *str, int base) {
         }
 
         while (value > 0) {
-                unsigned int digit = value % (unsigned int)base;
-                buf[i++]           = (char)((digit < 10) ? ('0' + digit) : ('a' + digit - 10));
+                u64 digit = value % (unsigned int)base;
+                buf[i++]  = (char)((digit < 10) ? ('0' + digit) : ('a' + digit - 10));
                 value /= (unsigned int)base;
         }
